Can now change name by pressing enter, centered a little more

+++ core/src/me/shreyasr/arrow/entity/attack/AttackBow.java
@@ -89,7 +89,7 @@ public class AttackBow implements Attack {
         if (timer<0 && !target.isEmpty()) {
             double startDir = target.getDir() - spread * spreadShots /2;
             for(int i=0;i< spreadShots;i++) {
-                Game.projectiles.add(new Projectile(new PolarVelocity(startDir + i * spread, speed),
+                Game.projectiles.add(new Projectile(new PolarVelocity((float)(startDir + i * spread), (float) speed),
                         Game.player.pos, "arrow"));
             }
             if (shotCounter <= 0) {
